requirements.txt
quart
quart-cors
quickchart.io